// src/types/Id.ts
export enum Id {
    none = "none",

    user1 = "user1",
    user2 = "user2",

    zeldaBreathOfTheWild = "zeldaBreathOfTheWild",
    marioKart8Deluxe = "marioKart8Deluxe",
    animalCrossingNewHorizons = "animalCrossingNewHorizons",
    humanFallFlat = "humanFallFlat",
    superSmashBrothersUltimate = "superSmashBrothersUltimate",
    gangBeasts = "gangBeasts",
    allApps = "allApps",

    news = "news",
    eShop = "eShop",
    album = "album",
    controller = "controller",
    settings = "settings",
    standby = "standby",
}